(feat) client-side prediction (no reconciliation)
Immediately update the position on the client-side after sending the
message to the server.

This provides the user with a smoother experience by eliminating the
delay of: sending an update to the server, waiting for the server to
process the update, and receiving the update from the server.

To the user, it now looks like the game immediately responds to their
input, instead of after a delay.

However, without reconciliation, the user will experience some jolting,
especially when making rapid changes in direction, because the server
updates immediately "jolt" the player back to the position that it was
in N frames ago.

So, next up, reconciliation.