feat: entity interpolation
Clients can employ prediction for their user-controlled entity to enjoy
smooth movement despite network delay and infrequent updates. But for
other entities, you must interpolate (or extrapolate, but that is less
frequently used).

Interpolation uses real entity data sent from the server to the client,
but the key is that it is played back to the user at a delay. When the
client doesn't have a server update at exactly the delay-time it wants
to render at, it can linearly interpolate, or 'lerp', the entity's
position.

If game updates are sent infrequently enough (like only 1-3 times per
second), interpolation could result in smooth-but-strange movement such
as the entity moving slowly between two points despite the entity
actually having made several rapid changes in direction.